Skeleton of standalone template migrator/verifier.
Added some utility classes for creating sane default log4net settings for
console and daemon applications, with support for variable verbosity of
console output. Handles STDERR and STDOUT 'properly', meaning that Error
or higher log levels go to STDERR instead of STDOUT.

Logging configuration should be overridable in config files in a fairly
straightforward manner. The logger used for console output is logically
called 'Console', and its appenders are Console.STDOUT and Console.STDERR.
Runtime specification of verbosity should always override the configured
level of the logger (but not the appenders).

refs #14984.

git-svn-id: https://www.clifton.bluewire-technologies.com/svn/repo1/Epro/trunk/Common@52646 57411a5e-3c33-3248-98c9-2c94d72dd52e